+++ RELEASE_NOTES.md
@@ -1,5 +1,6 @@
 4.18.0-SNAPSHOT
 ---------------
+* GMB-1052 - EOF exception is received when the trigger file name is given as .INP
 * GMB-1051 - Conditional sweeper sweeps files marked as inactive
 
 4.17.2